sharelink: pagesFIT production hotfix
PagesFIT[1] (the default deploy environment for our script) returns HTTP
404 if the URL contains a querystring after subfolder path.
E.g.
 - https://pages.fit.cvut.cz/peckato1/parsingtbl?ahoj returns HTTP 404
 - https://pages.fit.cvut.cz/peckato1/parsingtbl/?ahoj returns HTTP 404
 - https://pages.fit.cvut.cz/peckato1/parsingtbl/index.html?ahoj works!

So, to allow the link copy feature to work in this environment, let's
check if we run in a path that ends with /peckato1/parsingtbl(/)? and
then add the 'index.html' component into the copied URL.

I hope this will get fixed soon. The ticket against ICT FIT is opened
under #31823. This commit should be reverted when the fix lands.

Bug: [FIT #31823]